Update comets
COMETS can now be used in a modular fashion. The user can choose to run the entire process from any point or can decide to run only one process at a time. In addition, it is now possible to choose to create phyloseq_tables for All taxonomic groups or for Prokaryotes.